+++ Assets/Player/Script/IWeapon.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+public interface IWeapon
+{
+    void Fire();
+    void MakeKinematic(bool kine=true);
+    Material GetMaterial();
+    void SetMaterial(Material material);
+    void Reload();
+    void SetLock(bool enabled);
+    bool IsLocked();
+}