+++ ChangeLog.md
@@ -1,3 +1,54 @@
+## 0.2.1
+* Update navmesh after door is open
+* Gnome find new hide after hit
+* Immediate destroy door
+
+## 0.2.0-alpha.27
+* Fixed navmesh
+
+## 0.2.0-alpha.26
+* Added shotgun bullet
+* Added Exit
+* Fixed bullet life
+* Gnome flees on gun fire
+* Updated navmesh
+
+## 0.2.0-alpha.25
+* Added New walls
+* Added data center door
+* Mounted safe to datacenter wall
+* Moved shotgun to safe
+
+## 0.2.0-alpha.24
+* Added shotgun shells
+
+## 0.2.0-alpha.23
+* Added shotgun safe
+* Added breakable glass
+* Added test Exit
+* Added floppy Sound
+* Added glass Sound
+* Added shotgun sounds
+
+## 0.2.0-alpha.19
+* Added Shotgun
+* Added Shotgun animation
+* Added Floppy weapon
+* new weapon mechanism
+
+## 0.2.0-alpha.16
+* Updated earthquake code
+* Added prequake to earthquake
+* Fixed flurescent lamp placement
+* Changed intensity on flurescent lamp
+
+## 0.2.0-alpha.6
+* Player is nevmesh obstacle
+* Fixed Cable script again
+* Balanced volume
+* Added projectile impact sound
+* Changed gnome sound trigger
+
 ## 0.1.7
 * Code refactoring
 * Added help text to login screen